requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pandas==2.2.3
pika==1.3.2
celery==5.4.0
django-celery-results==2.5.1
rabbitmq==0.2.0
